fix(capabilities): isolate report contract from mutation cohort
